requirements.txt
django==3.1.7
django-rest-framework==0.1.0
djoser==2.1.0
psycopg2-binary==2.8.6
python-dotenv==0.15.0
django-username-email==2.4.1
pillow==8.1.2